requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.35.0
jinja2==3.1.3
python-dotenv==1.1.1
pydantic==2.6.4
pydantic_core==2.16.3
pyTelegramBotAPI==4.28.0
alpaca-trade-api==3.2.0
yfinance==0.2.38
msgpack>=1.0.8
openai==1.30.2
pandas>=2.3,<2.4
numpy>=1.26,<2.4
httpx==0.27.0
